% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_group_combinations.R
\name{calc_group_combinations}
\alias{calc_group_combinations}
\title{calc_group_combinations}
\usage{
calc_group_combinations(group, time_window = 10)
}
\arguments{
\item{group}{a data.table group with the columns:
content_id, id_user, timestamp_share}

\item{time_window}{Default to 10.
A integer with the time window of coordination.}
}
\value{
A data.table with coordinated contents.
Columns: content_id, id_user, content_id_y, id_user_y, time_delta
}
\description{
calc_group_combinations
}
