% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_twitter.R
\name{dt_unnest_wider}
\alias{dt_unnest_wider}
\title{dt_unnest_wider}
\usage{
dt_unnest_wider(dt, columns, keep = FALSE, simplify = TRUE)
}
\arguments{
\item{dt}{A data.table as input}

\item{columns}{the name(s) of the column(s) to be unnested (as string)}

\item{keep}{Defaults to TRUE. Keep the column to be unnested.}
}
\description{
Utility function for data.table.
Unnest a column into wide format. Takes the first row as reference!
Tries to emulate the functionality of tidyr::unnest_wider()
}
\details{
Current implementation probably very slow.
}
