% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_tweets_json}
\alias{load_tweets_json}
\title{load_tweets_json}
\usage{
load_tweets_json(data_dir)
}
\arguments{
\item{data_dir}{string that leads to the directory containing JSON files}
}
\value{
a data.table with all tweets loaded
}
\description{
Load Tweets stored in JSON files. Wrapper of the function \link[RcppSimdJson]{fload}
}
