% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_twitter.R
\name{preprocess_tweets}
\alias{preprocess_tweets}
\title{preprocess_tweets}
\usage{
preprocess_tweets(tweets)
}
\arguments{
\item{tweets}{a data.table with the (nested) columns: "entities", "public_metrics", "tweet_id", "created_at"}
}
\value{
a list with 5 data.tables: tweets (contains all tweets and their meta-data), referenced (information on referenced tweets), urls (all urls mentioned in tweets), mentions (other users mentioned in tweets), hashtags (hashtags mentioned in tweets)
}
\description{
Reformat nested twitter data (retrieved from Twitter v2 API); spreads out columns and reformats nested data.frame (data.table) to unnested sub-data.frames. All output is in long-format
}
