% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_twitter.R
\name{preprocess_twitter_users}
\alias{preprocess_twitter_users}
\title{preprocess_twitter_users}
\usage{
preprocess_twitter_users(users)
}
\arguments{
\item{users}{a data.table with unformatted (nested user data)}
}
\value{
a data.table with reformatted user data
}
\description{
Reformat nested twitter userdata (retrieved from Twitter v2 API); spreads out columns and reformats nested data.frame (data.table) to long format
}
