% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_tweets.r
\name{reshape_tweets}
\alias{reshape_tweets}
\title{reshape_tweets}
\usage{
reshape_tweets(tweets, intent = c("retweets", "hashtags", "urls"))
}
\arguments{
\item{tweets}{a pre-processed list of twitter data (output of preprocess_twitter())}

\item{intent}{the desired intent for analysis.}
}
\value{
a reshaped data.table
}
\description{
Reshape twitter data for coordination detection
}
